Move dom functions into props
