+++ apps/src/sites/studio/pages/projects/index.js
@@ -6,8 +6,10 @@ import PublicGallery from '@cdo/apps/templates/projects/PublicGallery';
 import experiments from '@cdo/apps/util/experiments';
 import ProjectHeader from '@cdo/apps/templates/projects/ProjectHeader';
 import i18n from '@cdo/locale';
+import {Galleries} from '@cdo/apps/templates/projects/GallerySwitcher';
 
 const MAX_PROJECTS_PER_CATEGORY = 100;
+const isPublic = window.location.pathname.startsWith('/projects/public');
 
 $(document).ready(() => {
   if (experiments.isEnabled('publicGallery')) {
@@ -16,7 +18,7 @@ $(document).ready(() => {
     $('#angular-my-projects-wrapper').attr('data-isPublicGalleryEnabled', 'true');
 
     const projectsHeader = document.getElementById('projects-header');
-    ReactDOM.render(<ProjectHeader />, projectsHeader);
+    ReactDOM.render(<ProjectHeader showGallery={showGallery} isPublic={isPublic}/>, projectsHeader);
 
     $.ajax({
       method: 'GET',
@@ -30,6 +32,17 @@ $(document).ready(() => {
     });
   }
 });
+
+function showGallery(gallery) {
+  updateLocation(gallery);
+  $('#angular-my-projects-wrapper').toggle(gallery === Galleries.PRIVATE);
+  $('#public-gallery-wrapper').toggle(gallery === Galleries.PUBLIC);
+}
+
+function updateLocation(gallery) {
+  const path = (gallery === Galleries.PUBLIC) ? '/projects/public' : '/projects';
+  window.history.pushState(null, document.title, path);
+}
 
 function onShowConfirmPublishDialog(callback) {
   const publishConfirm = document.getElementById('publish-confirm');
