+++ apps/src/templates/projects/ProjectHeader.jsx
@@ -4,33 +4,25 @@ import i18n from "@cdo/locale";
 import GallerySwitcher, {Galleries} from '@cdo/apps/templates/projects/GallerySwitcher';
 import HeadingBanner from '@cdo/apps/templates/HeadingBanner';
 
-const isPublic = window.location.pathname.startsWith('/projects/public');
-
 const ProjectHeader = React.createClass({
+  propTypes: {
+    showGallery: React.PropTypes.func.isRequired,
+    isPublic: React.PropTypes.bool
+  },
+
   render() {
     return (
-        <div>
-          <HeadingBanner
-            headingText={i18n.projectGalleryHeader()}
-          />
-          <GallerySwitcher
-            initialGallery={isPublic ? Galleries.PUBLIC : Galleries.PRIVATE}
-            showGallery={showGallery}
-          />
-        </div>
+      <div>
+        <HeadingBanner
+          headingText={i18n.projectGalleryHeader()}
+        />
+        <GallerySwitcher
+          initialGallery={this.props.isPublic ? Galleries.PUBLIC : Galleries.PRIVATE}
+          showGallery={this.props.showGallery}
+        />
+      </div>
     );
   }
 });
-
-function showGallery(gallery) {
-  updateLocation(gallery);
-  $('#angular-my-projects-wrapper').toggle(gallery === Galleries.PRIVATE);
-  $('#public-gallery-wrapper').toggle(gallery === Galleries.PUBLIC);
-}
-
-function updateLocation(gallery) {
-  const path = (gallery === Galleries.PUBLIC) ? '/projects/public' : '/projects';
-  window.history.pushState(null, document.title, path);
-}
 
 export default ProjectHeader;
