+++ dashboard/app/views/layouts/_header.html.haml
@@ -99,7 +99,7 @@
 - if show_header2
   - if request.path == "/home" || request.path == "/courses"
     #header-banner{style: 'background-image: url("/shared/images/banners/banner-test3.jpg")'}
-  - if request.path == "/projects"
+  - if request.path == "/projects" || request.path == "/projects/public"
     #header-banner{style: 'background-image: url("/shared/images/banners/banner_default.jpg")'}
 
 - if script_level
